requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
async-timeout==4.0.3
attrs==25.3.0
blinker==1.9.0
blis==1.2.0
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
confection==0.1.5
contourpy==1.3.1
cycler==0.12.1
cymem==2.0.11
distro==1.9.0
exceptiongroup==1.2.2
fonttools==4.56.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
ko_core_news_lg @ https://github.com/explosion/spacy-models/releases/download/ko_core_news_lg-3.8.0/ko_core_news_lg-3.8.0-py3-none-any.whl
langchain==0.3.20
langchain-core==0.3.45
langchain-openai==0.3.9
langchain-text-splitters==0.3.6
langcodes==3.5.0
langsmith==0.3.15
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
murmurhash==1.0.12
narwhals==1.31.0
numpy==2.2.4
openai==1.66.3
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pillow==11.1.0
preshed==3.0.9
protobuf==5.29.3
pyarrow==19.0.1
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.23.1
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.39
srsly==2.5.1
streamlit==1.43.2
tenacity==9.0.0
thinc==8.3.4
tiktoken==0.9.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typer==0.15.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wasabi==1.1.3
weasel==0.4.1
wordcloud==1.9.4
wrapt==1.17.2
zstandard==0.23.0
